Refactor, Added ThreeDotsMenu.js
